requirements.txt
PyQt5
Pillow
openai
pytesseract
